test: add eip 7623 and fee tests + refactor
