+++ src/tests/mod.rs
@@ -1,0 +1,4 @@
+mod eip7623;
+mod eip7702;
+mod fees;
+mod l1_message;
